Clicking and Typing adds item to ToDoList
Error in that new item text was not showing up. Changed the item to not be constant to add the item the user was adding to the list by using the textController. Also switched functions of buttons so that the ok button added an item and the cancel button did nothingl